Allow to configure window options as functions
This adds the possibility for more dynamic window configuration. To do
so, it allows some window configuration options to be a function. If so,
it evaluates the function by calling it with the window number as first
parameter. The output of the function is taken as the actual
configuration value that is then passed to the Nui library.

In addition to this, it also makes the "relative" option of Nui
configurable. The default value reflects the former fixed constant.

A possible usage of this would be a configuration that sizes and
positions Navbuddy relative to the current window and with the same
width as the window.
Moreover this allows to easily require a minimum width of Navbuddy and
much more. It does so without putting the burden to the plugin, but
rather the experienced power user.

All changes are fully backwards compatible.